requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-cprofile-middleware==1.0.5
django-storages==1.12.3
djangorestframework==3.13.1
docutils==0.18.1
lazy-object-proxy==1.7.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
